Ajout de la gestion de la dernière activité des membres Discord et mise à jour des méthodes d'enregistrement et de suppression.

+++ src/database/Models/Utilisateurs_discord.ts
@@ -5,6 +5,7 @@ export class UtilisateursDiscord implements I_Utilisateurs_discord {
     discord_id!: string;
     pseudo_discord!: string;
     join_date_discord!: string;
+    dern_activite?: Date;
 
     constructor(discord_id: string, pseudo_discord: string, join_date_discord: string) {
         this.discord_id = discord_id;
@@ -36,11 +37,8 @@ export class UtilisateursDiscord implements I_Utilisateurs_discord {
             // Vérifie si le membre est déjà enregistré
             const utilisateurDiscordDb = await db.select(UtilisateursDiscord.getTableName(), [utilisateurDiscord.discord_id]);
 
-            // Si le membre est déjà enregistré, met à jour la date de join et le pseudo
-            if (Array.isArray(utilisateurDiscordDb.results) && utilisateurDiscordDb.results.length > 0) {
-                await db.update(UtilisateursDiscord.getTableName(), utilisateurDiscord, `discord_id = ${utilisateurDiscord.discord_id}`);
-                return;
-            }
+            // Si le membre est déjà enregistré, ne rien faire
+            if (Array.isArray(utilisateurDiscordDb.results) && utilisateurDiscordDb.results.length > 0) return;
 
             // Enregistre le membre dans la base de données
             await db.insert(UtilisateursDiscord.getTableName(), utilisateurDiscord);
@@ -54,10 +52,36 @@ export class UtilisateursDiscord implements I_Utilisateurs_discord {
         const db = new Database();
 
         try {
+
+            // Vérifie si le membre est déjà enregistré
+            const utilisateurDiscordDb = await db.select(UtilisateursDiscord.getTableName(), [discord_id]);
+
+            // Si le membre n'est pas déjà enregistré, ne rien faire
+            if (Array.isArray(utilisateurDiscordDb.results) && utilisateurDiscordDb.results.length === 0) return;
+
             // Supprime le membre de la base de données
             await db.delete(UtilisateursDiscord.getTableName(), [discord_id]);
         } catch (error) {
             console.error('❌ Erreur lors de la suppression de l\'utilisateur : ', error);
+        }
+    }
+
+    // Enregistrement de la dernière activité du membre
+    static async update_dern_activite(utilisateurDiscord: UtilisateursDiscord) {
+        const db = new Database();
+
+        try {
+            // Vérifie si le membre est déjà enregistré
+            const utilisateurDiscordDb = await db.select(UtilisateursDiscord.getTableName(), [utilisateurDiscord.discord_id]);
+
+            // Si le membre n'est pas déjà enregistré, ne rien faire
+            if (Array.isArray(utilisateurDiscordDb.results) && utilisateurDiscordDb.results.length === 0) return;
+
+            // Met à jour la date de join
+            await db.update(UtilisateursDiscord.getTableName(), { dern_activite: utilisateurDiscord.dern_activite }, `discord_id = ${utilisateurDiscord.discord_id}`);
+
+        } catch (error) {
+            console.error('❌ Erreur lors de la mise à jour de la date de join : ', error);
         }
     }
